Found the bug!
Turns out I was reading the feature from the CSV file on the wrong order.

Now that the feature vector has the expected format, an "ideal-ish" trader, which buys for the minimum price of the day and sells for the maximum price of the day, consistently profits 0.4% / day.

The learning algorithm, on the other hand, still converges to 0 :(